server/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.3
Flask-MySQL==1.4.0
docxtpl==0.6.3
pyjwt==1.7.1
pymysql==0.8.0
Flask-SQLAlchemy==2.3.0
mysqlclient==1.4.6
